Fix bug in ArrayDeque.resize, add random testing
With many random operations, it was possible for the position of the
data in the array to be such that the index of last was greater either
than the size of the Deque or greater than the size of the new array. My
original resize operation assumed the data would be split around the
ends of the array, and subtracted last from size to determine how far to
copy for the right side. This resulted in an attempt to call arraycopy
with a negative size. For the left side it copied from 0 to last, which
was fine most of the time just copied a lot of nulls, but if last was
larger than the new array it would attempt to call arraycopy with a size
larger than the target array. Both only affected shrinking.