Rewrite ArrayDeque.resize
Previous fix introduced another new bug in the autograder. I was having
trouble understanding my original implementation (and the correction was
just case-catching) so I rewrote it in a simpler way with the new
understanding of the problem I gained fixing the old bug. I have not yet
recreated the failing tests from the autograder but I needed to get a
grasp on what my code was really doing before I felt comfortable
approaching further bug tracking.

The original implementation assumed the list would always be split
around the ends of the array, and copied the front half to the front of
the new array and the back half to the back. It seemed to still work
when the deque shifted enough to become contiguous on accident, but
there were two bugs: the shift could cause the algorithm to try to copy
all of the front of a list, including null padding at the front, to an
array too small, or where the calculations of how far to copy the list
resulted in a negative size to copy. My original bug fix was to leave my
original implementation but add a case for when the deque shifted enough
to become contiguous in the array.

I have rewritten the code to instead always copy the list to a
contiguous section of the new array starting at index 0, with a
better-defined case checking for when the array is contiguous or not to
determine how to do that correctly. It currently passes all of my tests
but I need to add more tests for the ones failed in the autograder to be
sure it fixed those bugs as well.