Update ArrayDequeTest
Added some of the tests from the autograder. randomOperations now uses a
LinkedListDeque (known to pass all tests) as a control for the outcomes
of the operations. Also added get to randomOperations. Added a test for
adding and removing items over and over and one for a bad get attempt
(checks for AssertionError).