Some attempts to fix autograder but prob giving up
It seems like it works on my end but the Gradescope grader says my test
is failing on inputs where the test was supposed to pass with both the
test and the control result being null. I'm not really sure where to go
from here and I can't find any other solutions where people even
attempted this problem.